Implemented first pass at client side and server side authentication.
Still need to persist the logged in user between calls. Also need to have smarter authorization scheme. It is now just an array of roles.

We also need to tie express routes to route handler and inject proper controllers

Also need to figure out how to tie in the passport authentication with our concept of controllers.